tools/verify_nemo/requirements.txt
torch
nemo_toolkit[asr]
librosa
